Week 2
- Array and method examples

- Deliverable 2
Create a C# Console application that uses two methods:

Method 1: Randomly populate an array with integers numbers between 10 and 50. The array length is a parameter for the method.
Method 2: Return the summation of all the elements of an array (the array is the parameter for this method).
The program has to follow the instructions below:

Ask the user to enter an integer number between 5 and 15.
Use a try/catch block to validate the user’s entry.
Use a conditional statement to check the user’s entry is between 5 and 15.
Use Method 1 to randomly populates an array with a length equal to the user’s input.
Display in the console all the elements the in the array.
Use Method 2 to sum all the elements in the array from step 4.
Display in the console the value from step 6.