Week 3
Deliverable 3 - Python Data Structures

Step (1): Import the pandas, numpy, and seaborn libraries. (didn't use numpy so removed it)

Step (2): Use the code line below to load the "iris" dataset and store it as a data frame.

Step (3): Drop the "species" column in the iris data frame.

Step (4): Add two more columns to the iris data frame, "sepal_sum" and "petal_sum". The "sepal_sum" is the sum of the values from the "sepal_length" and "sepal_width" columns. On the other hand, The "petal_sum" is the sum of the values from the "petal_length" and "petal_width" columns.

Step (5): Create a new data frame with the summary stats for the iris dataset. For this summary, compute each column's mean, standard deviation, and minimum and maximum values.