Week 4
Deliverable 4 - Python Programming

Method 1: Randomly populate a list with integers numbers between 10 and 50. The list length is a parameter for the method.
Method 2: Return the summation of all the elements of a list (the list is the parameter for this method).

Program Instructions
Ask the user to enter an integer number between 5 and 15.
Use a conditional statement to check the user’s entry is between 5 and 15.
Use Method 1 to randomly populates a list with a length equal to the user’s entry.
Display in the console all the elements the in the list.
Use Method 2 to sum all the elements in the list from step 3.
Display in the console the value from step 5.

Optional: Use a loop to repeat steps 1 and 2 until the user provides a proper entry.